+++ src/graphql/event/rdp.rs
@@ -2,7 +2,9 @@ use async_graphql::{Context, ID, Object, Result, StringNumber};
 use chrono::{DateTime, Utc};
 use review_database::event as database;
 
-use super::{ThreatLevel, TriageScore, country_code, find_ip_customer, find_ip_network};
+use super::{
+    ThreatLevel, TriageScore, country_code, country_codes, find_ip_customer, find_ip_network,
+};
 use crate::graphql::{customer::Customer, network::Network, triage::ThreatCategory};
 
 #[allow(clippy::module_name_repetitions)]
@@ -37,8 +39,8 @@ impl RdpBruteForce {
     /// The two-letter country code of the originator IP address. `"XX"` if the
     /// location of the address is not known, and `"ZZ"` if the location
     /// database is unavailable.
-    async fn orig_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.orig_addr)
+    async fn orig_country(&self) -> &str {
+        country_code(&self.inner.orig_country_code)
     }
 
     /// Originator Customer
@@ -68,12 +70,8 @@ impl RdpBruteForce {
     /// The two-letter country code of the responder IP address. `"XX"` if the
     /// location of the address is not known, and `"ZZ"` if the location
     /// database is unavailable.
-    async fn resp_countries(&self, ctx: &Context<'_>) -> Vec<String> {
-        self.inner
-            .resp_addrs
-            .iter()
-            .map(|resp_addr| country_code(ctx, *resp_addr))
-            .collect()
+    async fn resp_countries(&self) -> Vec<&str> {
+        country_codes(&self.inner.resp_country_codes)
     }
 
     /// Responder Customer List
@@ -163,8 +161,8 @@ impl BlocklistRdp {
     /// The two-letter country code of the originator IP address. `"XX"` if the
     /// location of the address is not known, and `"ZZ"` if the location
     /// database is unavailable.
-    async fn orig_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.orig_addr)
+    async fn orig_country(&self) -> &str {
+        country_code(&self.inner.orig_country_code)
     }
 
     /// Originator Customer
@@ -195,8 +193,8 @@ impl BlocklistRdp {
     /// The two-letter country code of the responder IP address. `"XX"` if the
     /// location of the address is not known, and `"ZZ"` if the location
     /// database is unavailable.
-    async fn resp_country(&self, ctx: &Context<'_>) -> String {
-        country_code(ctx, self.inner.resp_addr)
+    async fn resp_country(&self) -> &str {
+        country_code(&self.inner.resp_country_code)
     }
 
     /// Responder Customer